+++ src/components/About/About.scss
@@ -7,7 +7,7 @@
     background-repeat: no-repeat;
     background-color: grey;
     background-size: cover;
-    background-position: center
+    background-position: center;
 }
 .paper {
     margin: 30px;
