+++ src/components/Home/Home.js
@@ -3,9 +3,7 @@ import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import {makeStyles} from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
-import img from '../images/nike.jpg'
 
 import Divider from '@material-ui/core/Divider';
 import Tab from './Tab';
